moved the old backend and prototypes a kafka producer for the twitter api
