--- MakeTwitterGreatAgain.service
@@ -1,15 +1,0 @@
-[Unit]
-Description=MakeTwitterGreatAgain
-After=network.target
-
-[Service]
-ExecStart=/var/www/MakeTwitterGreatAgain/core.js
-Restart=always
-User=nobody
-Group=nobody
-Environment=PATH=/usr/bin:/usr/local/bin
-Environment=NODE_ENV=production
-WorkingDirectory=/var/www/MakeTwitterGreatAgain
-
-[Install]
-WantedBy=multi-user.target
