+++ backend/training/index.js
@@ -1,0 +1,3 @@
+module.exports = {
+  init: require('./init')
+}
