+++ twitterapi/.gitignore
@@ -1,0 +1,5 @@
+.idea/
+__pycache__
+venv/
+config.json
+scrapers.json
